refactor(tts/styletts2): retire legacy 4-graph backend, ANE is the only path
The StyleTTS2-ANE 7-graph re-cut has been the shipping backend since
199d7ce49; this drops the legacy 4-graph orchestration that was kept
around for parity testing.

- Delete StyleTTS2Manager + StyleTTS2Synthesizer (legacy pipeline).
- Delete StyleTTS2AneParityTests (cross-backend test no longer applies).
- Strip 4-graph CoreML loaders from StyleTTS2ModelStore; keep only the
  shared-asset surface (vocab, config, voices, repo root).
- Trim ModelNames.StyleTTS2 to vocab/config/voices only.
- Drop textPredictorBuckets / decoderBuckets / diffusionBucket from
  StyleTTS2Constants (now ANE-only).
- Remove .styleTts2 case from TtsBackend; route legacy CLI aliases
  ("styletts2", "style-tts2", "styletts", "style") to .styleTts2Ane.
- Remove --variant flag from `fluidaudiocli styletts2`; remove runLegacy
  + runStyleTts2 from the benchmark harness.
- Port tokenizeWithReport + voicesRepoRoot accessors to
  StyleTTS2AneManager so the CLI no longer needs the legacy manager.
- Update Architecture.md / Benchmarks.md / StyleTTS2Ane.md to reflect
  the single-backend reality.

Repo.styleTts2 is preserved — the HF repo still hosts the shared vocab
+ config + voice presets used by the ANE backend.
