+++ Sources/FluidAudioCLI/Commands/TtsBenchmarkCommand.swift
@@ -277,12 +277,6 @@ public enum TtsBenchmarkCommand {
                     voice: voice,
                     preset: preset, outputJson: outputJson, audioDir: audioDir,
                     asrChoice: asrChoice)
-            case .styleTts2:
-                try await runStyleTts2(
-                    phrases: phrases, corpusLabel: corpusLabel,
-                    voicePath: voice,
-                    preset: preset, outputJson: outputJson, audioDir: audioDir,
-                    asrChoice: asrChoice)
             case .styleTts2Ane:
                 try await runStyleTts2Ane(
                     phrases: phrases, corpusLabel: corpusLabel,
@@ -668,69 +662,6 @@ public enum TtsBenchmarkCommand {
         }
     }
 
-    // MARK: - StyleTTS2 driver
-
-    private static func runStyleTts2(
-        phrases: [(category: String, text: String)],
-        corpusLabel: String,
-        voicePath: String?,
-        preset: TtsComputeUnitPreset,
-        outputJson: String?,
-        audioDir: String?,
-        asrChoice: AsrChoice
-    ) async throws {
-        guard let voicePath, !voicePath.isEmpty else {
-            logger.error(
-                "StyleTTS2 requires --voice <path/to/ref_s.bin> "
-                    + "(256 fp32 LE blob from mobius-styletts2/scripts/06_dump_ref_s.py)")
-            exit(1)
-        }
-        let voiceURL = resolveURL(voicePath, isDirectory: false)
-        let voiceLabel = voiceURL.deletingPathExtension().lastPathComponent
-
-        // StyleTTS2 doesn't expose a compute-units knob today; --compute-units
-        // is accepted for parity with other backends but only labels the run.
-        let manager = StyleTTS2Manager()
-
-        let coldStart = Date()
-        try await manager.initialize()
-        let coldStartS = Date().timeIntervalSince(coldStart)
-        logger.info(String(format: "Cold start (initialize): %.2fs", coldStartS))
-
-        let firstStart = Date()
-        _ = try await manager.synthesizeSamples(
-            text: "Initialization warm-up.", voiceStyleURL: voiceURL, randomSeed: 42)
-        let firstSynthMs = Date().timeIntervalSince(firstStart) * 1000
-        logger.info(String(format: "First synth: %.0f ms", firstSynthMs))
-
-        try await runPhraseLoop(
-            backendId: "styletts2",
-            voiceLabel: voiceLabel,
-            corpusLabel: corpusLabel,
-            phrases: phrases,
-            preset: preset,
-            coldStartS: coldStartS,
-            firstSynthMs: firstSynthMs,
-            outputJson: outputJson,
-            audioDir: audioDir,
-            asrChoice: asrChoice,
-            extraSummary: ["voice": voiceLabel]
-        ) { text in
-            let t0 = Date()
-            let result = try await manager.synthesizeSamples(
-                text: text, voiceStyleURL: voiceURL, randomSeed: 42)
-            let synthMs = Date().timeIntervalSince(t0) * 1000
-            return BackendPhraseSample(
-                synthMs: synthMs,
-                ttftMs: synthMs,
-                samples: result.samples,
-                sampleRate: result.sampleRate,
-                stageMs: [:],
-                extraFields: [:]
-            )
-        }
-    }
-
     // MARK: - StyleTTS2-ANE driver (7-graph re-cut)
 
     private static func runStyleTts2Ane(
@@ -1070,7 +1001,6 @@ public enum TtsBenchmarkCommand {
         case pocketTts
         case magpie
         case cosyVoice3
-        case styleTts2
         case styleTts2Ane
 
         var defaultCorpus: String {
@@ -1093,10 +1023,9 @@ public enum TtsBenchmarkCommand {
             return .magpie
         case "cosyvoice3", "cosyvoice", "cosy":
             return .cosyVoice3
-        case "styletts2-ane", "styletts2ane", "styletts2_ane", "style-ane", "s2-ane":
+        case "styletts2-ane", "styletts2ane", "styletts2_ane", "style-ane", "s2-ane",
+            "styletts2", "style-tts2", "styletts", "style":
             return .styleTts2Ane
-        case "styletts2", "style-tts2", "styletts", "style":
-            return .styleTts2
         default:
             logger.warning("Unknown backend '\(name)' — defaulting to kokoro-ane")
             return .kokoroAne
